protocol_lib/requirements.txt
crcmod

pyserial